% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummExpDR.R
\name{addColData}
\alias{addColData}
\title{add colData}
\usage{
addColData(x, value, col_name = NULL)
}
\arguments{
\item{value}{DataFrame or coercible thereof, or vector. if vector, col_name must be specified}
}
\description{
add colData
}
