% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\name{annotate_probes}
\alias{annotate_probes}
\title{build annotations}
\usage{
annotate_probes(
  illumina_anno,
  granges_prep_list,
  prefix,
  coord_column = "pos",
  chr_column = "chr",
  n_cores = 1,
  pbar = TRUE
)
}
\arguments{
\item{illumina_anno}{= data frame from illumina EPIC array manifest file, must have MAPINFO column containing cpg coordinates}

\item{granges_prep_list}{= list of granges_prep objects}

\item{prefix}{= column name prefix for annotations}

\item{coord_column}{= column for coordinates of methylation sites}

\item{chr_column}{= column with chromosomes that probes are annotated to}

\item{pbar}{= whether or not to add progress bar}
}
\value{
modified illumina_anno data frame with columns added for gene body annotations (semicolon delimited list of genes, '' if empty set),
promoter annotations (semicolon delimited list of genes, '' if empty set), and enhancer annotations (semicolon delimited list of start and end points
of enhancers).
}
\description{
build annotations
}
