% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlap.R
\name{calc_overlap}
\alias{calc_overlap}
\title{Calculate Overlap between character all vectors in list(s)}
\usage{
calc_overlap(
  list1,
  list2 = NULL,
  metric = "num.ovr",
  master_set = NULL,
  n_cores = 1,
  pbar = TRUE
)
}
\arguments{
\item{metric}{= num.ovr for number of overlapping elements in character vectors a nd b,
jaccard for jaccard coefficient, ovr.coef for intersect(a,b)/union(a,b),
fisher.p for -log10 p.value of fisher's exact test}

\item{master_set}{= master geneset used for fisher's exact test. set1 and set2 must be subset of this set.
only used if fisher.p option selected}

\item{n_cores}{= number of cores to use in parallel. If NULL, use detected cores on machine - 1.
if n_cores set to 1, a serial job is done}

\item{pbar}{whether or not to use progress bar}
}
\value{
If 2 lists provided, and m x n matrix M where m is number of
elements in list1, and n is number of elements in list2, and M[i,j] is
calculated overlap between sets list1[[i]] and list2[[j]]. If 1 list provided,
an nxn matrix where n is length of list1, M[i,j] is calculated overlap of
list1 i and list1 j
}
\description{
Given two lists of character vectors, calculate overlap for each
combination of character vectors from first and second list, return as matrix. If one list is
provided, calculate overlap for each combination of character vectors within given list.
}
