% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{calc_sim_mats}
\alias{calc_sim_mats}
\title{Calculate pairwise ARI matrices}
\usage{
calc_sim_mats(object)
}
\arguments{
\item{object}{iter_clust_obj output by do_clust_}
}
\value{

}
\description{
Calculate pairwise ARI matrices for resampled clustering
solutions, permuted labels, and, if multiple funs for clustering
on full dataset, for solutions found for full dataset_
}
