% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_utils.R
\name{calculate_probe_avg}
\alias{calculate_probe_avg}
\title{Calculate Genewise Average Of Probe Feature}
\usage{
calculate_probe_avg(inp_data, gene_probe_mapping, n_cores = 1, pbar = TRUE)
}
\arguments{
\item{inp_data}{= input data matrix, rows corresponding to probes, columns corresponding to features}

\item{gene_probe_mapping}{= gene to probe mapping. must be dataframe containing genes in rows and column 'mapped_probes' of semicolon delimited strings
denoting EPIC probes}

\item{n_cores}{= number of cores to use}
}
\value{
list with matrix of gene-wise averages, genes in rows, features in columns, and a second entry with number of probes per gene.
Note that genes without any probes are excluded from this matrix and probe count entry
}
\description{
Calculate Genewise Average Of Probe Feature
}
