% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\name{check_granges_prep}
\alias{check_granges_prep}
\title{check validity of granges_prep object}
\usage{
check_granges_prep(granges_prep)
}
\arguments{
\item{granges_prep}{granges_prep object}
}
\description{
check validity of granges_prep object
}
