% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_utils.R
\name{check_min_probes}
\alias{check_min_probes}
\title{Get check if given gene has min number of probes}
\usage{
check_min_probes(min_probes, data_mat, gene_probe_mapping)
}
\arguments{
\item{min_probes}{= minimum number of probes for a gene}

\item{data_mat}{= nxm feature x sample matrix}

\item{gene_probe_mapping}{= gene to probe mapping. must be dataframe containing genes in rows and column 'mapped_probes' of semicolon delimited strings
denoting EPIC probes}
}
\description{
Get check if given gene has min number of probes
}
