% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{clust_once}
\alias{clust_once}
\title{Run one iteration of clustering}
\usage{
clust_once(inp_mat, algorithm, k, seed)
}
\arguments{
\item{inp_mat}{matrix with numeric entries_ Affinity matrix if algorithm is spectral_kernlab or spectral_SNFtool,
sample x feature matrix normalized by columns if kmeans_ see \code{\link{make_inp_mat}}}

\item{algorithm}{character_ spectral_kernlab runs spectral clustering as implemented in kernlab package,
spectral_SNFtool runs spectral clustering as implemented in SNFtool package, kmeans runs
kmeans clustering as implemented in \code{\link{stats::kmeans}} package}
}
\value{
numeric_ cluster labels in the order of samples in the input matrix
}
\description{
Run one iteration of clustering
}
