% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{compare_full_rs}
\alias{compare_full_rs}
\title{Compare clustering solutions on full data to clustering
solutions on resampled data to determine similarity}
\usage{
compare_full_rs(object)
}
\arguments{
\item{object}{iter_clust_obj}
}
\value{
iter_clust_obj with compar_mat_full_rs assigned as a matrix
of pairwise similarities between all clustering solutions on full
dataset (rows) with all resampled clustering solutions, and with
soln_metadata slot having a column added for median ARI compared to
resampled solutions
}
\description{
Compare clustering solutions on full data to clustering
solutions on resampled data to determine similarity
}
