% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{compare_rs_perm}
\alias{compare_rs_perm}
\title{Compare distributions of pairwise ARI for resampled cluster solutions
and permuted labels}
\usage{
compare_rs_perm(object)
}
\arguments{
\item{object}{iter_clust_obj}
}
\value{
iter_clust_obj with sim_rs_v_perm filled in with
result of one sided wilcoxon ranksum test_ Note that
the alternative hypothesis is that the distribution of
ARI values for the resampled clustering solutions is
higher than that of the the permuted labels
}
\description{
Compare distributions of pairwise ARI for resampled cluster solutions
and permuted labels
}
