% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{corplot_2_var}
\alias{corplot_2_var}
\title{2 variable correlation plot}
\usage{
corplot_2_var(
  df,
  var1,
  var2,
  color_by = NULL,
  shape_by = NULL,
  label = NULL,
  method = "pearson",
  filter_by = NULL,
  filter_class = NULL,
  ...
)
}
\arguments{
\item{df}{= data.frame}

\item{var1}{= x axis variable}

\item{var2}{= y axis variable}

\item{color_by}{= variable to color points by}

\item{shape_by}{= variable to shape points by}

\item{label}{= variable to label samples by}

\item{method}{method for correlation}

\item{filter_by}{= subset data by a particular categorical variable}

\item{filter_class}{= set of allowed classes for subset}
}
\description{
2 variable correlation plot
}
\details{
note that correlation is calculated within filtered subset of data
}
