% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\name{create_granges_prep}
\alias{create_granges_prep}
\title{create granges_prep}
\usage{
create_granges_prep(granges_obj, operation, analysis_name, bp = 200L)
}
\arguments{
\item{granges_obj}{= GRanges object}

\item{operation}{= operation to perform (string, one of 'in' or 'upstream')}

\item{analysis_name}{= name for granges analysis}

\item{bp}{= number of basepairs upstream}
}
\description{
create granges_prep object
}
