% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{create_multi_k_clust}
\alias{create_multi_k_clust}
\title{create a multi_k_clust_object}
\usage{
create_multi_k_clust()
}
\value{
\code{\link{multi_k_clust_object}}
}
\description{
create a multi_k_clust_object
}
