% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummExpDR.R
\name{create_SummExpDR}
\alias{create_SummExpDR}
\title{Create SummExpDR object}
\usage{
create_SummExpDR(summ_exp)
}
\arguments{
\item{summ_exp}{SummarizedExperiment object}
}
\description{
Create SummExpDR object
}
