% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierDetection.R
\name{detect_outliers}
\alias{detect_outliers}
\title{Detect outliers}
\usage{
detect_outliers(object, assay_use = 1)
}
\arguments{
\item{object}{SummExpDR class object, SummarizedExperiment, or mxn numeric
data matrix.}

\item{assay_use}{which assay to use for outlier detection}
}
\value{
if a SummExpDR, a SummExpDR with metadata updated
with an outliers column. if a SummarizedExperiment, a SummarizedExperiment
with colData updated with an outliers column. if mxn matrix, a logical
vector indicating outlier status
}
\description{
Detect outliers
}
