% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Analysis.R
\name{do_fisher_multi}
\alias{do_fisher_multi}
\title{Fisher's Exact Test on all pathways in gmt given query geneset}
\usage{
do_fisher_multi(
  query,
  pathway_list,
  master_set,
  FDR = 0.1,
  n_cores = 1,
  pbar = FALSE
)
}
\arguments{
\item{query}{character vector with query feature set}

\item{pathway_list}{list of pathways of interest}

\item{FDR}{FDR to filter at}

\item{n_cores}{number of cores to use for fisher's exact test}

\item{pbar}{whether or not to use a progressbar}
}
\description{
Fisher's Exact Test on all pathways in gmt given query geneset
}
