% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\name{.annotate_1_probe}
\alias{.annotate_1_probe}
\title{Helper function used for annotate_probes}
\usage{
.annotate_1_probe(illumina_coords, probe_idx, granges_prep_list)
}
\arguments{
\item{illumina_coords}{= data frame containing probe coordinates (column 1) and corresponding chromosomes (column 2)}

\item{probe_idx}{= probe index (integer to index illumina_coords by row)}

\item{transcript_grange}{= GenomicRanges object with transcript coordinates}

\item{enh_grange}{= GenomicRanges objec with enhancer coordinates}
}
\value{
1 row dataframe with a column for each granges_prep object in granges_prep_list
}
\description{
Helper function used for annotate_probes
}
