% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\name{.check_isin}
\alias{.check_isin}
\title{Check if a probe is within a set of genomic regions}
\usage{
.check_isin(coord, chr, regions)
}
\arguments{
\item{coord}{= query coordinate}

\item{chr}{= chromosome indexed}

\item{regions}{= GRanges object with start + end points}
}
\value{
logical vector indicating which regions the coordinate lies within
}
\description{
Check if a probe is within a set of genomic regions
}
