% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\name{.check_upstream}
\alias{.check_upstream}
\title{Check if probes within within specified distance upstream of coordinate}
\usage{
.check_upstream(coord, chr, regions, bp = 200)
}
\arguments{
\item{coord}{= query coordinate}

\item{chr}{= chromosome indexed}

\item{regions}{= GenomicRanges object}

\item{bp}{= filter for probes within this distance of TSS}
}
\value{
logical vector indicating for which genes the coord is upstream of
}
\description{
Check if probes within within specified distance upstream of coordinate
}
