% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\docType{class}
\name{DR-class}
\alias{DR-class}
\title{Dimensionality Reduction Container}
\description{
inherits from SummarizedExperiment. assay data = dimensionality reduction
coordinates, rows correspond to PCs, cols correspond to samples
}
\section{Slots}{

\describe{
\item{\code{sourceAssay}}{assay used to create Dimensionality Reduction object.
This comes in handy for varimax rotation, where knowledge of which assay
used for PCA is necessary to recompute reduced dims.}
}}

