% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Analysis.R
\name{enrich_GSEA}
\alias{enrich_GSEA}
\title{Run GseaPreranked from R via system call}
\usage{
enrich_GSEA(
  gsea_rank_list_path,
  gsea_analysis_name = "analysis",
  gsea_rpt_label = "report",
  gsea_jar_path = file.path("~/GSEA/gsea-3.0.jar"),
  gsea_out = NULL,
  gsea_gmx = NULL,
  gsea_num_permutations = 1000,
  gsea_min_gs_size = 15,
  gsea_max_gs_size = 200,
  plot_top = 20,
  seed = 12345,
  mem = "16G"
)
}
\arguments{
\item{gsea_rank_list_path}{character, length 1. path to ranked list file}

\item{gsea_analysis_name}{character, length 1. name for gsea analysis}

\item{gsea_rpt_label}{character, length 1. report label}

\item{gsea_jar_path}{path to GSEA jar file}

\item{gsea_out}{directory under which GSEA results will be output. GSEA makes a directory
under this one to contain the GSEA output}

\item{gsea_gmx}{path to gmt file}

\item{gsea_num_permutations}{number of permutations to run for GSEA}

\item{gsea_min_gs_size}{minimum geneset size for GSEA}

\item{gsea_max_gs_size}{maximum geneset size for GSEA}

\item{plot_top}{generate plots for top (plot_top) positively and negatively enriched pathways}

\item{seed}{numeric, length 1. seed to use}

\item{mem}{character, length 1. memory (in GB) to allot to GSEA. E.g. 16G for 16 GB}
}
\value{
exit status of system call to GseaPreranked , numeric of length 1.
}
\description{
Run GseaPreranked from R via system call
}
\details{
assumes you have GSEA version 3.0 .jar file in your filesystem
}
