% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\docType{class}
\name{FactorizedDR-class}
\alias{FactorizedDR-class}
\title{class for linear dimensionality reduction that
can have data reconstructed from factors}
\description{
class for linear dimensionality reduction that
can have data reconstructed from factors
}
