% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummExpDR.R
\name{fetchData}
\alias{fetchData}
\title{Interface to fetch data from object}
\usage{
fetchData(
  x,
  varsFetch,
  redDimKeys = NULL,
  assayKey = NULL,
  mode = "sample_wise"
)
}
\arguments{
\item{x}{SummExpDR object}

\item{varsFetch}{variables to pull out}

\item{redDimKeys}{keys for reduced dims to pull out}

\item{assayKey}{assay to pull from summExp object}

\item{mode}{whether to pull data in a sample x feature format or feature x feature_info format
for sample x feature (sample_wises), first pull colData of summ exp object, then join with embeddings from redDimKeys
and with assay data from assay keys. for feature x feat info, pull out rowData, then any loadings for reduced dim keys}
}
\description{
Interface to fetch data from object
}
