% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{filter_NA}
\alias{filter_NA}
\title{Filter NA values from deseq2 result}
\usage{
filter_NA(DESEQ2.result)
}
\arguments{
\item{DESEQ2.result}{DESeqResults object}
}
\value{
DESeqResults object with NA values removed
}
\description{
Filter NA values from deseq2 result
}
