% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Analysis.R
\name{fisher.test.sets}
\alias{fisher.test.sets}
\title{Fisher's Exact Test, with set input}
\usage{
fisher.test.sets(set1, set2, master_set)
}
\arguments{
\item{set1}{character vector containing a set of labels}

\item{set2}{character vector containing a set of labels}

\item{master_set}{set of which set1 and set2 are a subset of and from which they were sampled}
}
\value{
the result of fisher.test run for a contingency table made given the input sets
}
\description{
computes fisher's exact test for association between set1 and set2
given a master set from which set1 and set 2 were drawn.
}
