% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_utils.R
\name{get_mapped_probes}
\alias{get_mapped_probes}
\title{Map gene to probes}
\usage{
get_mapped_probes(gene, gene_probe_mapping)
}
\arguments{
\item{gene}{= gene name}

\item{gene_probe_mapping}{= gene to probe mapping. must be dataframe containing genes in rows and column 'mapped_probes' of semicolon delimited strings
denoting EPIC probes}
}
\description{
Map gene to probes
}
