% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{get_qc_k}
\alias{get_qc_k}
\title{Report metrics for qualitiy control for single k}
\usage{
get_qc_k(
  iter_clust_obj,
  thresh_pval = 0.05,
  thresh_ARI = 0.9,
  thresh_proportion = 0.75
)
}
\arguments{
\item{iter_clust_obj}{object of class iter_clust_obj, produced from
\code{\link{run_iter_clust_functions}}}

\item{thresh_pval}{if wilcoxon ranksum test comparing ARI for resampled clusterings
to that for permuted labels gives p_value <= this argument, consider
the result significant}

\item{thresh_ARI}{threshold ARI for pairwise comparisons of resampling based
clustering solutions, and pairwise comparisons of reported clustering solution
to resampling based clustering solutions
run_iter_clust_functions}

\item{thresh_proportion}{threshold proportion of}
}
\value{
data.frame with 1 row, and the following columns:
- is_signif: logical
- prop_rs_v_rs: numeric
- prop_rpt_v_rs: numeric
- meets_qc: logical
see details for more info
}
\description{
Report metrics for qualitiy control for single k
}
\details{
the following QC metrics will be reported:
statistical significance for wilcoxon rank sum test comparing pairwise ARI
for resampled clustering solutions and those obtained for permuted labels,
determined with specified p_value threshold; p_value for said test (is_signif);
proportion of pairwise ARI for resampling based clustering solutions with
ARI above specified threshold (prop_rs_v_rs);
proportion of pairwise ARI in comparisons of reported solution (for given k)
to resampling based clusters (prop_rpt_v_rs);
does proportion of pairwise ARI values for the reampled clustering comparisons
and, and that calculated for the reported solution x resampled solutions comparisons,
exceed a specified threshold proportion (meets_qc);
These values are all stored in teh data.frame that is returned
}
