% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{get_sim_dist}
\alias{get_sim_dist}
\title{return vector of pairwise similarities}
\usage{
get_sim_dist(sim_mat)
}
\arguments{
\item{sim_mat}{output of function get_sim_mat}
}
\value{
numeric vector of all non-self pairwise ARI
}
\description{
return vector of pairwise similarities
}
\details{
returns all pairwise ARI from sim_mat,
i_e_ sim_mat[i,j] for all i, j, j > i (avoid counting twice)
}
