% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{get_sim_mat}
\alias{get_sim_mat}
\title{Compute pairwise adjusted rand index matrix}
\usage{
get_sim_mat(M)
}
\arguments{
\item{M}{matrix of clustering solutions, each row being a different
solution_ must have > 1 row}
}
\value{
sim_mat matrix of pairwise similarities for input clustering
solutions_ Let sim_mat be the output matrix_ then sim_mat[i,j] is the ARI
between solutions at rows i and j of input matrix x_ If solutions i and j
do not have a completely overlapping set of samples (due to resampling
applied in iter_clust), then ARI is calculated between all samples in the
intersection of samples between the two solutions
}
\description{
compute pairwise adjusted rand index
(ARI) between all clustering solutions in input matrix
}
