% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{get_soln_iter_clust}
\alias{get_soln_iter_clust}
\title{Retrieve Solution From iter_clust_obj}
\usage{
get_soln_iter_clust(object, get_best = T, soln_num = NULL)
}
\arguments{
\item{object}{iter_clust_object}

\item{get_best}{logical of length 1_ if true, retrieve solution
automatically picked to be reported_ if flase, must set soln_num
to an integer value}

\item{soln_num}{numeric of length 1_ which solution to pick, if
get_best set to false_ must be set to an integer value within 1:
number of solutions if get_best set to false}
}
\value{
named numeric vector of cluster labels for solution
selected
}
\description{
Retrieve Solution From iter_clust_obj
}
