% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{get_soln_k}
\alias{get_soln_k}
\title{Retrieve a clustering solution out of a multi_k_clust object}
\usage{
get_soln_k(
  object,
  pick_best_valid = T,
  k = NULL,
  analysis_use = 1,
  add_2_metadata = F
)
}
\arguments{
\item{object}{multi_k_clust object output by \code{\link{get_qc_all_k}} or
\code{\link{run_multi_k_functions}}, or a SummExpDR object output by
\code{\link{run_multi_k_functions}}}

\item{pick_best_valid}{pick best solution for highest value k that passes all qc
metrics as described in \code{\link{get_qc_all_k}}}

\item{k}{NULL, or integer specifying which k to use. if clustering results at
given k did not pass qc results, generate a warning but still return result. If
specified, return solution reported for clustering analysis with k clusters,}

\item{analysis_use}{name of clustering analysis, or an integer specifying index
of cluster analysis performed. Specifies which multi_k_clust object to retrieve
from a SummExpDR object class object}

\item{add_2_metadata}{add cluster labels to metadata. new metadata column will be
named according to the name of the clustering analysis and the value k (# of
clusters). E.g. clust.analysis_k3 for clustering analysis 'clust.analysis',
best solution for k3
(reported by corresponding iter_clust_obj)}
}
\value{
named integer vector of clustering labels, or if object is of class SummExpDR
and add_2_metadata = TRUE, a SummExpDR with cluster labels added to metadata
}
\description{
retrieve best clustering solution from a given value k,
by default, the highest value k that passes 'qc' metrics.
}
