% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{get_soln_metadata}
\alias{get_soln_metadata}
\title{Retrieve Metadata for Unique Solutions}
\usage{
get_soln_metadata(object)
}
\arguments{
\item{object}{iter_clust_obj}
}
\value{
data.frame, metadata for unique solutions_
For more info, see documentation for class iter_clust_obj
}
\description{
Retrieve Metadata for Unique Solutions
}
