% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummExpDR.R
\name{getAnalyses}
\alias{getAnalyses}
\title{Getter for analyses slot}
\usage{
getAnalyses(x, key)
}
\arguments{
\item{x}{SummExpDR object}

\item{key}{name of analysis to retrieve}
}
\description{
Getter for analyses slot
}
