% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{getEmbeddings}
\alias{getEmbeddings}
\title{Get Embeddings}
\usage{
getEmbeddings(x, key = NULL, rows = NULL, cols = NULL)
}
\arguments{
\item{x}{DR (Dimensionality Reduction) class object or SummExpDR object}

\item{key}{key if x is SummExpDR object}

\item{rows}{= rows to subset (integer or logical indices or character vector)}

\item{cols}{= cols to subset (integer or logical indices or character vector)}
}
\description{
Get Embeddings
}
