% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{getLoadings}
\alias{getLoadings}
\title{Get Loadings from Linear Dimensionality Reduction Object}
\usage{
getLoadings(x, key = NULL, red_dims = NULL, features = NULL)
}
\arguments{
\item{x}{LinearDR object or}

\item{key}{key for dimensionality reduction}

\item{red_dims}{index for reduced dims. Ideally should be character.}

\item{features}{names of features to use (e.g. gene names), or NULL for all features. If not null should ideally be character}
}
\description{
Get Loadings from Linear Dimensionality Reduction Object
}
