% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\docType{class}
\name{granges_prep-class}
\alias{granges_prep-class}
\title{Container For GRanges Object + Operation}
\description{
Container For GRanges Object + Operation
}
\section{Slots}{

\describe{
\item{\code{granges_obj}}{= GRanges object, with names set}

\item{\code{operation}}{= operation to perform, 1 of in/upstream}

\item{\code{analysis_name}}{= name for granges analysis}

\item{\code{bp}}{= integer: number of base pairs upstream if checking upstream}
}}

