% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiExpIntegration.R
\name{imputeExpData}
\alias{imputeExpData}
\title{Impute Missing Data in multiExp object}
\usage{
imputeExpData(x, ...)
}
\arguments{
\item{x}{multiExp object}

\item{...}{other args for sklearn.impute.KNNImputer}
}
\description{
Impute Missing Data in multiExp object
}
