% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\docType{class}
\name{iter_clust_obj-class}
\alias{iter_clust_obj-class}
\title{Object to run iterative clustering/resampling procedure for single k}
\description{
Object to run iterative clustering/resampling procedure for single k
}
\section{Slots}{

\describe{
\item{\code{clusts_full}}{NULL or matrix of clustering solutions obtained on
full dataset_}

\item{\code{clusts_rs}}{NULL or matrix of clustering solutions obtained on
dataset when resampled in each iteration}

\item{\code{clusts_rs_perm}}{NULL or matrix resulting from permuting rows of
clusts_rs}

\item{\code{sim_mat_full}}{NULL or pairwise ARI matrix calculated by get_sim_mat on
clusts_full}

\item{\code{sim_mat_rs}}{NULL or pairwise ARI_matrix calculated by get_sim_mat
on clusts_rs}

\item{\code{sim_mat_rs_perm}}{NULL or pairwise ARI matrix calculated by
get_sim_mat on clusts_rs_perm}

\item{\code{unique_solns}}{NULL or matrix containing unique solutions found in
clusts_rs by unique_solutions method}

\item{\code{soln_metadata}}{NULL or data.frame containing metadata for unique_solns
solutions_ Metadata include proportion of solutions in clusts_full
accounted for by given unique solution, median similarity (ARI) to solutions
in clusts_rs, median absolute deviation (MAD) for similarity to solutions
in cluts_rs (derived from compar_mat_full_rs), and whether this cluster
solution is to be reported_ Currently, the decision to report is done first
by maximum median similarity to resampled solutions, and in case of ties,
by highest proportion of solutions accounted for in clusts_full_ In
case of further ties, pick the first solution out of the set of
solutions for which we need to break ties_}

\item{\code{compar_mat_full_rs}}{NULL, or matrix M with M[i,j] = ARI
between solution i (from unique solutions) and solution j
(from clusts_rs)}

\item{\code{sim_rs_v_perm}}{the results of a
wilcoxon ranksum test comparing distributions of pairwise ARI for
cluster labels obtained from resampled clustering and those obtained
from permuting resampled clustering based labels_}
}}

