% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{KNN_impute}
\alias{KNN_impute}
\title{Wrapper around sklearn.immpute.KNNImputer}
\usage{
KNN_impute(x, ...)
}
\arguments{
\item{x}{data, features in rows, samples in columns}

\item{...}{other args for sklearn.impute.KNNImputer()}
}
\description{
Wrapper around sklearn.immpute.KNNImputer
}
