% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Imputation.R
\name{KNN_scale_impute}
\alias{KNN_scale_impute}
\title{Perform imputation with rescaling of data}
\usage{
KNN_scale_impute(x, scale_factors = NULL, ...)
}
\arguments{
\item{x}{data, features in rows, samples in columns}

\item{scale_factors}{if NULL, set to 1/stddev for all feats. otherwise, must be numeric
of length nrow(x), and names of scale factors must correspond to names of features}

\item{...}{other args for sklearn.impute.KNNImputer()}
}
\description{
Data is scaled by multiplying each variable by corresponding entry of scale_factors (scale_factors must)
}
