% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Analysis.R
\name{load_gmt}
\alias{load_gmt}
\title{Load gmt file}
\usage{
load_gmt(
  gmt_file,
  min_genes = 15,
  max_genes = 200,
  master_set = NULL,
  n_cores = 1,
  pbar = FALSE
)
}
\arguments{
\item{gmt_file}{gmt_file to load in}

\item{min_genes}{minimum genes per pathway. Note that min filter is applied after filtering for genes in master set}

\item{max_genes}{maximum genes per pathway. Note that max filter is applied after filtering for genes in master set}

\item{master_set}{master gene set, of which all pathways must be a subset in the result}

\item{n_cores}{number of cores to use for filtering}

\item{pbar}{whether or not to use a progressbar}
}
\description{
Load gmt file
}
\details{
in addition to filtering for genes (strings) in master set, filters out empty characters (\'\').
}
