% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DNAm_anno.R
\name{make_gene_probe_mapping}
\alias{make_gene_probe_mapping}
\title{Make gene to probe mapping}
\usage{
make_gene_probe_mapping(illumina_anno, mapping_col, n_cores = 1, pbar = TRUE)
}
\arguments{
\item{illumina_anno}{= data.frame or DataFrame of probe metadata, containing
a mapping of probes to genes for a particular type of annotation, e.g. gene body, enhancer, promoter}

\item{mapping_col}{= name of column containing mapping. column contains character vector of semicolon delimited
strings denoting genes each probe maps to}

\item{n_cores}{= number of cores to use}

\item{pbar}{= whether to include progress bar}
}
\description{
Make gene to probe mapping
}
