% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{make_input_mat}
\alias{make_input_mat}
\title{make input matrix for clustering}
\usage{
make_input_mat(
  x,
  std_norm,
  mode = "snf_affi",
  affi_K = 20,
  sigma = 0.5,
  as_kernMat = F
)
}
\arguments{
\item{x}{matrix, with samples in rows and features in columns_
Note that in future we may try to incorporate SNF into this (i_e_
multiple data matrices input for SNF pipeline)}

\item{std_norm}{logical, length 1_ standard normalize data (normalize columns)?}

\item{mode}{mode by which to construct affinity matrix_ if snf_affi,
then construct an affinity matrix using dist2 and affinityMatrix functions,
passing euclidean distance matrix to affinityMatrix function_
If std_norm is true, then data standard normalized prior to affinity matrix
construction_
If mxn, then return mxn matrix, standard normalized if std_norm is true_
Currently, only snf_affi and mxn supported
(i_e_ for hierarchical clustering)_}

\item{affi_K}{integer of length 1_ K parameter for \code{\link{SNFtool::affinityMatrix}}}

\item{sigma}{numeric of length 1_ sigma parameter for affinityMatrix}

\item{as_kernMat}{logical, length 1_ coerce output matrix to class \code{\link{kernlab::kernelMatrix}}?}
}
\value{
an affinity matrix
}
\description{
make input matrix for clustering
}
