% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{make_iter_clust_obj}
\alias{make_iter_clust_obj}
\title{Constructor Function for iter_clust_obj}
\usage{
make_iter_clust_obj()
}
\value{
new instance of iter_clust_obj
}
\description{
Constructor Function for iter_clust_obj
}
