% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\docType{class}
\name{multi_k_clust-class}
\alias{multi_k_clust-class}
\title{object to run clustering analyses for multiple values k}
\description{
object to run clustering analyses for multiple values k
}
\section{Slots}{

\describe{
\item{\code{single_k_analyses}}{either NULL or a named list of iter_clust_obj
objects. Naming convetion for list: single_k_analyses[['k']] =
iter_clust_obj run for k clusters}

\item{\code{report}}{data.frame with rows corresponding to clustering analyses
at k = (row #), and with first (n -1) columns corresponding to variables
of interest used to determine the value k selected.
Columns include:
- is.significant = are pairwise ARI for resampled clusters significantly
higher than would be expected under a null hypothesis of permuted labels?
- rs.ARI.gt.thresh = proportion of pairwise ARI greater than threshold set
- full.v.rs.ARI.gt.thresh = proportion of pairwise ARI between reported
solution on full data and resampled solutions exceeding threshold
-report.k = report solution for this value k?}
}}

