% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Analysis.R
\name{pathways_to_enrichment_map_fmt}
\alias{pathways_to_enrichment_map_fmt}
\title{Convenience function to convert results of do_fisher_multi to a generic EM file}
\usage{
pathways_to_enrichment_map_fmt(df, fpath)
}
\arguments{
\item{df}{dataframe. should be output by do_fisher_multi}

\item{fpath}{filepath}
}
\value{
doesn't return anything, just writes Enrichment Map generic formatted file
}
\description{
Convenience function to convert results of do_fisher_multi to a generic EM file
}
\details{
See https://enrichmentmap.readthedocs.io/en/docs-3.1/FileFormats.html for details on
enrichment map generic file format.
}
