% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{perm_labels}
\alias{perm_labels}
\title{permute cluster labels in matrix}
\usage{
perm_labels(M, base_seed = 2)
}
\arguments{
\item{M}{matrix output by iter_clust}

\item{base_seed}{smallest seed used_ seeds used
for permutations range from base_seed to base_seed
plus nrow(M) (i_e_ number of iterations of clustering)
minus 1}
}
\value{
matrix, with each row permuted
}
\description{
permute cluster labels in matrix
}
