% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{pick_soln}
\alias{pick_soln}
\title{Report solution to use}
\usage{
pick_soln(object)
}
\arguments{
\item{object}{iter_clust_obj}
}
\value{
iter_clust_obj with soln_metadata slot having a column added
for median ARI compared to resampled solutions, and a column added
denoting whether or not the solution is the one to be reported
for this iter_clust_obj_ Note that in case of algorithms that
only have 1 clustering performed on full data (e_g_ SNFtool
implementation of spectral clustering), this degenerates
to picking the only solution_
}
\description{
Report solution to use
}
\details{
if more than one unique solution to pick from,
pick the solution with the highest median ARI when compared to
resampled solutions_ In case of ties, pick solution with highest
proportion of clustering solutions in full dataset_ In cases
of further ties, pick the first solution our of the set of
the ties for proportion
}
