% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plot_sim_dist}
\alias{plot_sim_dist}
\title{plot distributions of pairwise ARI}
\usage{
plot_sim_dist(
  object,
  clust_analysis = 1,
  mode = "resample.vs.perm",
  file = "pairwise_ARI.png",
  save_file = F
)
}
\arguments{
\item{clust_analysis}{clustering analysis to select out of
clust.results slot. Can be an integer indicating position
in list, or alternatively the name assigned to said analysis}

\item{mode}{if resample.vs.perm (default, only option), plots distributions
of pairwise ARI for resampled clustering solutions vs permuted
labels. if full.vs.rs, pairwise ARI from best solution at each k
to the resampled clustering solutions (this is to assess if
chosen solution is basically the same as solutions found via resampling).}

\item{file}{filepath}

\item{save_file}{save file? if TRUE, save file. if FALSE, a boxplot is produced and sent
to the default device for plots}
}
\description{
plot distributions of pairwise ARI
}
