% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{correlation plot method,}
\usage{
plotCorrelation(x, feat1, feat2, color_by, key = NULL, assay = NULL, ...)
}
\arguments{
\item{x}{SummExpDR object}

\item{feat1}{x variable to plot}

\item{feat2}{y variable to plot}

\item{key}{key for reduced dims to use}

\item{assay}{required if pulling ouot feature values for assay data}

\item{...}{other arguments to corplot_2_var}
}
\description{
correlation plot method,
}
