% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotDR}
\alias{plotDR}
\title{Plot Reduced Dims}
\usage{
plotDR(x, dim1, dim2, key, color_by, label = NULL, assay = NULL, ...)
}
\arguments{
\item{x}{SummExpDR object}

\item{key}{key for reduced dims to use}

\item{assay}{required if pulling ouot feature values for assay data}

\item{...}{other arguments to scatter_plot, with var1 and var2 overridden by dim1 and dim2}
}
\description{
Plot Reduced Dims
}
