% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotVarianceExplained}
\alias{plotVarianceExplained}
\title{Plot variance explained across all features and per datatype}
\usage{
plotVarianceExplained(x, key, dims_use = NULL, do_print = TRUE)
}
\arguments{
\item{x}{multiExp object}

\item{key}{key for reduced dims}

\item{do_print}{logical, do print stuff}
}
\description{
Plot variance explained across all features and per datatype
}
