% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{rank_and_order}
\alias{rank_and_order}
\title{Rank and order genes according to Log Fold Change and adjusted p-value}
\usage{
rank_and_order(
  LFC_padj_df,
  write.rankfile = T,
  output_dir = getwd(),
  fname = "ranked_list.rnk"
)
}
\arguments{
\item{LFC_padj_df}{data.frame with log fold change RNA exp in column 1, FDR (padj) in second}

\item{write.rankfile}{write a rankfile?}

\item{output_dir}{filepath to directory where you want to write rankfile}

\item{fname}{filename, to be appended to out_}
}
\value{
data frame with genes in first column, score in second
}
\description{
Rank and order genes according to Log Fold Change and adjusted p-value
}
\details{
scores are calculated as -log10(padj)*sign(LFC). To account for Inf and -Inf values,
any scores with abs(score) == Inf are converted to sign(LFC)*(max(abs(score.i)) + runif(0,1)),
s.t. score.i != inf. runif(0,1) random uniformly distributed variable, from range 0,1
}
