% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OutlierDetection.R
\name{remove_outliers}
\alias{remove_outliers}
\title{Filter outliers from data}
\usage{
remove_outliers(object)
}
\arguments{
\item{object}{either a SummarizedExperiment object or a SummExpDR class object}
}
\value{
object with outliers removed from data and metadata. if a SummExpDR,
all slots except expt.data are destroyed as output is a new SummExpDR with
subsetted experimental data and no entries for clustering analyses + differential
phenotype analysis
}
\description{
remove outliers from data
}
