% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Analysis.R
\name{retrieve_gsea}
\alias{retrieve_gsea}
\title{Extract results tables for positively and negatively enriched pathways}
\usage{
retrieve_gsea(gsea_results_dir)
}
\arguments{
\item{gsea_results_dir}{character, length 1. path to GSEA results.}
}
\value{
list with 2 entries: na_pos_report = contents of gsea_report_for_na_pos\[0-9\]*.xls,
na_neg_report = contents of gsea_report_for_na_neg\[0-9\]*.xls
}
\description{
Extract results tables for positively and negatively enriched pathways
}
