% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{run_clust_multi_k}
\alias{run_clust_multi_k}
\title{run clustering/evaluation workflow for multiple k}
\usage{
run_clust_multi_k(
  object,
  inp_data = NULL,
  k_use = 2:10,
  num_iter = 200,
  std_norm = NULL,
  algorithm = "spectral_kernlab",
  pct_resample = 0.8,
  affi_K = 20,
  sigma = 0.5,
  base_seed = 2,
  ...
)
}
\arguments{
\item{object}{multi_k_clust object}

\item{k_use}{integer vector, length >= 1. values k to perform
clustering for}

\item{num_iter}{numeric, length 1_ number of iterations of clustering_
Note that if we are running a deterministic clustering algorithm
without resampling, this argument automatically gets set to 1}

\item{std_norm}{standard normalize data prior to clustering?
and is set prior to sampling the data_}

\item{algorithm}{character_ spectral_kernlab runs spectral clustering as implemented in kernlab package,
spectral_SNFtool runs spectral clustering as implemented in SNFtool package, kmeans runs
kmeans clustering as implemented in \code{\link{stats::kmeans}} package}

\item{pct_resample}{percentage of data to resample_ Specify as number
greater than 0 and less than 1}

\item{affi_K}{numeric, length 1_ K value for affinityMatrix_
used if algorithm = 'spectral_SNFtool' or 'spectral_kernlab'}

\item{sigma}{numeric of length 1_ sigma value for affinityMatrix_
used if algorithm = 'spectral_kernlab' or 'spectral_SNFtool'}

\item{base_seed}{integer of length 1_ lowest seed in range
base_seed:(base_seed + num_iter - 1) used for resampling,
and in case of clustering with restarts, clustering_ For resampling,
in each iteration, the seed used is incremented up in this range,
and before clustering the seed is reset to the current seed in the range
(i_e_ in a given iteration the seed set for resampling and clustering are
the same)}

\item{...}{other args for iter_clust}
}
\value{
multi_k_clust object with single.k.analysis slot assigned
as a list of iter_clust_obj objects for which clustering and
evaluation of full data and resampled solutions has been done.
Each element of the list is named according to the value k that
was specified for the numnber of clusters.
for more dtails, see run_iter_clust_functions.
}
\description{
run clustering/evaluation workflow for multiple k
}
