% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{run_DESEQ2}
\alias{run_DESEQ2}
\title{Run DESEQ2 on count data}
\usage{
run_DESEQ2(
  expt_data,
  class_type_use,
  class1,
  class2 = NULL,
  n.cores = 1,
  assay_use = 1,
  covariates = character(0),
  remove_low_cts = F,
  rowsum.thresh = 5
)
}
\arguments{
\item{expt_data}{SummarizedExperiment object with count data.}

\item{class_type_use}{class denoted in metadata (must be one of metadata columns)}

\item{class1}{value for first class, which will be 'numerator' for log fold change}

\item{class2}{will be 'denominator' for log fold change. if set to NULL,
defined as the complement of class1 (i.e. all other classes), and called 'all.others'}

\item{n.cores}{number of cores to use for DESeq2 differential expression}

\item{assay_use}{which assay to use from expt_data. should indicate counts}

\item{covariates}{to use in linear model or GLM if using 'voom', 'limma', or
'DESEQ2' pipelines. Must be character. If length 0, design matrix consists only
of intercept and covariate of interest (class_type_use).
If length 1 or more, design matrix consists if intercept, covariate of interest,
ass less as additional covariates specified by this parameter}

\item{remove_low_cts}{remove low counts? If this is the case, remove all genes
with row sums less than rowsum.thresh}

\item{rowsum.thresh}{see remove_low_cts}
}
\value{
data frame of results with LFC defined by log2(c1/c2), where c1 and
c2 give the expression of class1 and class2, respectively. Uses BH corrected
p-values, 2 sided test (altHypothesis = 'greaterAbs'). Uses default
arguments values for DESeq2::\code{\link{results}}
}
\description{
Run DESEQ2 on count data
}
