% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{run_diff_exp}
\alias{run_diff_exp}
\title{Run differential expression analysis}
\usage{
run_diff_exp(
  expt_data,
  class_type_use,
  class1,
  class2 = NULL,
  covariates = character(0),
  assay_use = 1,
  pipeline = "voom",
  write_output = T,
  output_dir = "./",
  prefix = "",
  overwrite = F,
  ...
)
}
\arguments{
\item{expt_data}{SummarizedExperiment object with RNA-seq data or microarray
data, or SummExpDR with said SummarizedExperiment object in @summ_exp slot. If using
limma as selected differential expression pipeline, can handle DNAm m-values as well
or any data matrix where features are expected to have a Gaussian distribution}

\item{class_type_use}{class denoted in metadata (must be one of metadata columns)}

\item{class1}{value for first class, which will be 'numerator' for log fold change}

\item{class2}{will be 'denominator' for log fold change. if set to NULL,
defined as the complement of class1 (i.e. all other classes), and called 'all.others'}

\item{covariates}{to use in linear model or GLM if using 'voom', 'limma', or
'DESEQ2' pipelines. Must be character. If length 0, design matrix consists only
of intercept and covariate of interest (class_type_use).
If length 1 or more, design matrix consists if intercept, covariate of interest,
ass less as additional covariates specified by this parameter}

\item{assay_use}{which assay to pull out of SummarizedExperiment object.
should be 1 unless there are multiple assays and you want to get an assay that isn't
the first 1}

\item{pipeline}{one of voom, limma, or DESEQ2.}

\item{write_output}{write output to file, i.e. diff exp result, rankfile. files will be written
in subdirectory of output_dir}

\item{output_dir}{top output directory for}

\item{prefix}{specified prefix for rank files (i.e. filename = [prefix].rnk). If null,
defaults to [class1]_vs_[class2]}

\item{overwrite}{overwrite output files of prior run}

\item{...}{other arguments to pass to run_limma or run_DESEQ2}
}
\value{

}
\description{
Run differential expression analysis
}
\examples{

}
