% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gsva_scoring.R
\name{run_gsva}
\alias{run_gsva}
\title{Run GSVA}
\usage{
run_gsva(x, assay, genesets, expt_use = NULL, ...)
}
\arguments{
\item{assay}{assay to pull data from}

\item{genesets}{named list of genesets to pass to GSVA::gsva}

\item{expt_use}{experiment to use if this is a multiExp object}
}
\description{
Run GSVA
}
