% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{run_limma}
\alias{run_limma}
\title{Run limma or voom and limma}
\usage{
run_limma(
  expt_data,
  class_type_use,
  class1,
  class2,
  assay_use = 1,
  covariates = character(0),
  use_voom = F,
  remove_low_cts = T
)
}
\arguments{
\item{expt_data}{SummarizedExperiment object with RNA-seq data or microarray
data, or SummExpDR with said SummarizedExperiment object in @summ_exp slot. If using
limma as selected differential expression pipeline, can handle DNAm m-values as well
or any data matrix where features are expected to have a Gaussian distribution}

\item{class_type_use}{class denoted in metadata (must be one of metadata columns)}

\item{class1}{value for first class, which will be 'numerator' for log fold change}

\item{class2}{will be 'denominator' for log fold change. if set to NULL,
defined as the complement of class1 (i.e. all other classes), and called 'all.others'}

\item{assay_use}{assay to use from expt_data. defaults to first slot.}

\item{covariates}{covariates to include in linear model}

\item{use_voom}{use edgeR/voom/limma workflow if TRUE. If FALSE, use limma differential
expression experimental data provided. Should set to TRUE for RNA-seq data,
FALSE for log RMA values for microarray data}
}
\value{
differential expression results table as output by
toptable
}
\description{
Run limma or voom and limma
}
