% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{run_multi_k_functions}
\alias{run_multi_k_functions}
\title{Take Input Data through clustering with multiple values k}
\usage{
run_multi_k_functions(
  x,
  assay_use = 1,
  analysis_name = "clust.analysis",
  redDimKey = NULL,
  dims_use = NULL,
  std_norm = NULL,
  k_use = 2:10,
  num_iter = 200,
  algorithm = "spectral_kernlab",
  pct_resample = 0.8,
  affi_K = 20,
  sigma = 0.5,
  base_seed = 2,
  thresh_pval = 0.05,
  thresh_ARI = 0.9,
  thresh_proportion = 0.9,
  ...
)
}
\arguments{
\item{x}{either an mxn sample x feature matrix, or a SummExpDR object containing input data
or pca results}

\item{assay_use}{if SummExpDR object, which assay to use from expt.data slot}

\item{analysis_name}{name for clustering analysis. argument unused if x is a matrix}

\item{redDimKey}{if NULL, use coordinates}

\item{std_norm}{standard normalize data prior to clustering. for spectral clustering, this
means standard normalization prior to construciton of affinity matrix. If unspecified,
will exhibit following default behavior: For multi_clust_k_obj, defaults to TRUE;
for SummExpDR, if redDimKey is NULL, defaults to TRUE, if redDimKey is non-NULL,
defaults to FALSE}

\item{k_use}{values k to use for clustering}

\item{num_iter}{numeric, length 1_ number of iterations of clustering_
Note that if we are running a deterministic clustering algorithm
without resampling, this argument automatically gets set to 1}

\item{algorithm}{character_ spectral_kernlab runs spectral clustering as implemented in kernlab package,
spectral_SNFtool runs spectral clustering as implemented in SNFtool package, kmeans runs
kmeans clustering as implemented in \code{\link{stats::kmeans}} package}

\item{pct_resample}{percentage of data to resample_ Specify as number
greater than 0 and less than 1}

\item{affi_K}{numeric, length 1_ K value for affinityMatrix_
used if algorithm = 'spectral_SNFtool' or 'spectral_kernlab'}

\item{sigma}{numeric of length 1_ sigma value for affinityMatrix_
used if algorithm = 'spectral_kernlab' or 'spectral_SNFtool'}

\item{base_seed}{integer of length 1_ lowest seed in range
base_seed:(base_seed + num_iter - 1) used for resampling,
and in case of clustering with restarts, clustering_ For resampling,
in each iteration, the seed used is incremented up in this range,
and before clustering the seed is reset to the current seed in the range
(i_e_ in a given iteration the seed set for resampling and clustering are
the same)}

\item{...}{other args for iter_clust}
}
\description{
Take Input Data through clustering with multiple values k
}
