% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{runPCA}
\alias{runPCA}
\title{Run PCA on assay data}
\usage{
runPCA(x, i = NULL, suffix = NULL, std_norm = TRUE)
}
\arguments{
\item{x}{= SummExpDR, SummarizedExperiment, or matrix.
matrix must have the form n features x m samples}

\item{i}{= assay to pull}

\item{suffix}{suffix to add, if SummExpDR. By default, reduced dim key is just 'PCA', but with suffix added it's 'PCA{suffix}'}

\item{std_norm}{whether to standard normalize data}
}
\description{
Run PCA on assay data
}
