% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{runVarimax}
\alias{runVarimax}
\title{Run varimax rotation on a linear dimensionality reduction}
\usage{
runVarimax(x, key, dims_use = NULL, suffix = "", ...)
}
\arguments{
\item{x}{SummExpDR object}

\item{key}{key for LinearDR in ReducedDims to access}

\item{dims_use}{dims to use (character or integer/logical index, character preferred)}

\item{...}{other args to pass to stats::varimax}
}
\description{
Run varimax rotation on a linear dimensionality reduction
}
