% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiExpIntegration.R
\name{scaleExpData}
\alias{scaleExpData}
\title{Scale Experiment Data}
\usage{
scaleExpData(x, assay)
}
\arguments{
\item{x}{multiExp object}

\item{assay}{character, assay to pull out for scaling.}
}
\description{
Scale Experiment Data
}
