% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{screePlot}
\alias{screePlot}
\title{Scree Plot of Explained Variance}
\usage{
screePlot(x, key, feats_use = NULL, dims_use = NULL)
}
\arguments{
\item{x}{SummExpDR with FactorizedDR dimensionality reduction object computed}

\item{key}{string indicating which item from reducedDims to compute reconstruction accuracy for}

\item{feats_use}{integer vector or character vector indicating features to compute variance explained for. if NULL,
all features. Essentially you subset the data matrix and the reconstructed data matrix for the selected features,
then compute r-squared to compare those subset matrices.}

\item{dims_use}{dims to plot. if NULL plot all}
}
\description{
Scree Plot of Explained Variance
}
