% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Overlap.R
\name{set_overlap}
\alias{set_overlap}
\title{Calculate overlap between character vectors}
\usage{
set_overlap(set1, set2, metric = "num.ovr", master_set = NULL)
}
\arguments{
\item{set1}{= character vector}

\item{set2}{= character vector}

\item{metric}{= num.ovr for number of overlapping elements in character vectors a nd b,
jaccard for jaccard coefficient, ovr.coef for intersect(a,b)/union(a,b),
fisher.p for -log10 p.value of fisher's exact test}

\item{master_set}{= master geneset used for fisher's exact test. set1 and set2 must be subset of this set.
only used if fisher.p option selected}
}
\value{
calculated metric of overlap for set1 and set2
}
\description{
Calculate overlap between character vectors
}
