% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummExpDR.R
\name{setAnalyses}
\alias{setAnalyses}
\title{Setter for Analyses slot}
\usage{
setAnalyses(x, key, value)
}
\arguments{
\item{x}{SummExpDR}

\item{key}{character vector of length 1}
}
\description{
Setter for Analyses slot
}
