% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummExpDR.R
\name{setReducedDims}
\alias{setReducedDims}
\title{Setter for reducedDims slot}
\usage{
setReducedDims(x, key, value)
}
\arguments{
\item{x}{SummExpDR}

\item{key}{character vector of length 1}
}
\description{
Setter for reducedDims slot
}
