% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{setSourceAssay}
\alias{setSourceAssay}
\title{Set assay assay used to generate dim reduction in SummExpDR object}
\usage{
setSourceAssay(x, assay_name)
}
\description{
Set assay assay used to generate dim reduction in SummExpDR object
}
