% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{subs_by_class}
\alias{subs_by_class}
\title{Subset sample metadata by class}
\usage{
subs_by_class(metadata, class_type_use, class1, class2)
}
\arguments{
\item{metadata}{metadata}

\item{class_type_use}{class denoted in metadata (must be one of metadata columns)}

\item{class1}{value for first class, which will be 'numerator' for log fold change}

\item{class2}{will be 'denominator' for log fold change. if set to NULL,
defined as the complement of class1 (i.e. all other classes), and called 'all.others'}
}
\value{
metadata subsetted by class if 2 classes specified, or metadata
with column class_type_use changed to have classes [class1] and 'all.others'
}
\description{
Subset sample metadata by class
}
