% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummExpDR.R
\docType{class}
\name{SummExpDR-class}
\alias{SummExpDR-class}
\title{Wrapper around SummarizedExperiment object to contain Reduced Dims
as an additional slot}
\description{
Note that subsetting methods are not included as subsetting AFTER performing
dimensionality reduction typically means that you won't get the same coordinates
if dimensionality reduction is rerun.
}
\section{Slots}{

\describe{
\item{\code{summ_exp}}{SummarizedExperiment}

\item{\code{reducedDims}}{list of SummarizedExperiment objects, containing sample names (colData)}
}}

