% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IterativeClustering.R
\name{unique_solutions}
\alias{unique_solutions}
\title{Find unique solutions in full dataset clustering
and proportion of solutions accounted for}
\usage{
unique_solutions(object)
}
\arguments{
\item{object}{iter_clust_obj output by calc_sim_mats}
}
\value{
iter_clust_obj with unique solutions specified and
soln_metadata slot initialized as a data.frame containing
the proportion of solutions in the full dataset clusterings
accounted for by each unique solution_
}
\description{
Find unique solutions in full dataset clustering
and proportion of solutions accounted for
}
