% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{varianceExplained}
\alias{varianceExplained}
\title{method for calculating reconstruction accuracy}
\usage{
varianceExplained(x, key, dims_use = NULL, feats_use = NULL)
}
\arguments{
\item{x}{SummExpDR with FactorizedDR dimensionality reduction object computed}

\item{key}{string indicating which item from reducedDims to compute reconstruction accuracy for}

\item{dims_use}{integer vector or character vector indicating dimensions to use, or NULL for all dimensions}

\item{feats_use}{integer vector or character vector indicating features to compute variance explained for. if NULL,
all features. Essentially you subset the data matrix and the reconstructed data matrix for the selected features,
then compute r-squared to compare those subset matrices.}
}
\description{
Calculates reconstruction accuracy (r-squared) of matrix factorization w.r.t. input data for
fitting matrix factorization
}
