% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DifferentialExpression.R
\name{write_RNA_DE_output}
\alias{write_RNA_DE_output}
\title{Write ranked genelist and diff exp results to output files}
\usage{
write_RNA_DE_output(
  obj,
  result_name = NULL,
  top_output_dir = "./",
  overwrite = F
)
}
\arguments{
\item{obj}{either a su2c_obj class object or a list output by run_diff_exp}

\item{result_name}{name of result to retrieve from obj@diff.phen. NULL by default,
must be specified if obj is of class su2c_obj}

\item{top_output_dir}{top output directory below which a subdirectory containing
analysis output is kept}

\item{overwrite}{overwrite contents of output directory (subdirectory of top_output_dir).
In practice,  a system call is made to remove output directory and its contents
before making it again if output directory exists. If set to FALSE, will raise error if
directory already exists. If set to TRUE, and in the unlikely event the subdirectory created has
the same name as something you don't want to remove, provided that the user running the
process has the appropriate permissions, the directory will be removed.}
}
\value{
Nothing returned, just write out ranked genelist (see rank and order) and
differential expression results (see run_voom).
}
\description{
Write ranked genelist and diff exp results to output files
}
\details{
if list output by run_diff_exp, write out ranked genelist entry (rnk),
and write out differential expression entry (diff_exp)
}
