% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Pathway_Analysis.R
\name{do_fisher_multi}
\alias{do_fisher_multi}
\title{Fisher's Exact Test on all pathways in gmt given query geneset}
\usage{
do_fisher_multi(
  query,
  gmt_file,
  master_set,
  FDR = 0.1,
  n_cores = 1,
  pbar = FALSE
)
}
\arguments{
\item{query}{character vector with query feature set}

\item{gmt_file}{gmt file containing pathways of interest}

\item{FDR}{FDR to filter at}
}
\description{
Fisher's Exact Test on all pathways in gmt given query geneset
}
