% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MultiExpIntegration.R
\name{imputeExpData}
\alias{imputeExpData}
\title{Impute Missing Data in multiExp object}
\usage{
imputeExpData(x, assay = "stacked", use_num_feats = TRUE, ...)
}
\arguments{
\item{x}{multiExp object}

\item{assay}{assay for which data is to be imputed. by default it's stacked, the assay name for all of the input data stacked together}

\item{use_num_feats}{scale data in imputation space 1/(sd(feat)*sqrt(n_feats in datatype)), where feat is value of a feature, n_feats is
number of features in said datatype}

\item{...}{other args for sklearn.impute.KNNImputer}
}
\description{
Impute Missing Data in multiExp object
}
