% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{plotCorrelation}
\alias{plotCorrelation}
\title{correlation plot method,}
\usage{
plotCorrelation(
  x,
  feat1,
  feat2,
  color_by = NULL,
  shape_by = NULL,
  key = NULL,
  assay = NULL,
  method = "pearson",
  filter_by = NULL,
  filter_class = NULL,
  ...
)
}
\arguments{
\item{feat1}{x variable to plot}

\item{feat2}{y variable to plot}

\item{color_by}{= variable to color points by}

\item{shape_by}{= variable to shape points by}

\item{method}{method for correlation}

\item{filter_by}{= subset data by a particular categorical variable}

\item{filter_class}{= set of allowed classes for subset}

\item{...}{other arguments to corplot_2_var}
}
\description{
correlation plot method,
}
