% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DimReduction.R
\name{runPCA}
\alias{runPCA}
\title{#' method for calculating reconstruction accuracy
#'}
\usage{
runPCA(x, i = NULL, suffix = NULL, std_norm = TRUE)
}
\arguments{
\item{x}{= SummExpDR, SummarizedExperiment, or matrix.
matrix must have the form n features x m samples}

\item{i}{= assay to pull}

\item{suffix}{suffix to add, if SummExpDR. By default it's just \'PCA\'}

\item{std_norm}{whether to standard normalize data}
}
\description{
setGeneric('qualityMetric',  function(x, key, dims_use) standardGeneric("qualityMetric"))
}
\details{
setMethod('qualityMetric',
          signature = 'SummExpDR',
          function(x, key, dims_use) {
            FactorizedDR <- getReducedDims(x, key)
            assay_used <- FactorizedDR@sourceAssay
            input_data_SE <- getSummExp()
          })
Run PCA on assay data
}
