% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotFunctions.R
\name{scatter_plot}
\alias{scatter_plot}
\title{generic scatterplot function for data frame}
\usage{
scatter_plot(
  df,
  var1,
  var2,
  color_by = NULL,
  shape_by = NULL,
  label = NULL,
  filter_by = NULL,
  filter_class = NULL,
  pt_size = 0.5,
  alpha = 0.4,
  legend_pt_size = 20,
  xlim = NULL,
  ylim = NULL,
  lab_size = 1
)
}
\arguments{
\item{df}{= data.frame}

\item{var1}{= x axis variable}

\item{var2}{= y axis variable}

\item{color_by}{= variable to color points by}

\item{shape_by}{= variable to shape points by}

\item{label}{= variable to label samples by}

\item{filter_by}{= subset data by a particular categorical variable}

\item{filter_class}{= set of allowed classes for subset}
}
\description{
generic scatterplot function for data frame
}
